Factored out current state implementation and added abstraction layer
